Switch to fuse-based worker :sparkles:
Big change: use bb-worker's FUSE filesystem for runner build directory.

Requires worker to be a privileged container, so it can set up and share the FUSE
mount point to runner container.  Runner remains unprivileged, however (good).

Note FUSE mountpoint can't be in an emptyDir, and needs to be on a hostPath
filesystem.  See github comment and surrounding discussion:
https://github.com/kubernetes/kubernetes/issues/7890#issuecomment-1570824211